Perbaikan CAR

+++ app/Http/Controllers/Admin/KaryawanController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Karyawan;
+use App\Models\Role_Has_User;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
@@ -15,14 +16,16 @@ class KaryawanController extends Controller
     {
         $title = 'Karyawan';
         if ($request->has('search')) {
-            $dataKaryawan = User::whereIn('jabatan', ['Konsultan', 'Project Manager', 'Staff', 'Support Manager'])
-                ->where('nama', 'LIKE', '%' . $request->search . '%')
-                ->orderBy('nama', 'asc')
-                ->paginate(20);
+            $dataKaryawan = User::join('role_has_user', 'user.id', '=', 'role_has_user.fk_user')
+                ->where('role_has_user.fk_role', 2)
+                ->where('user.nama', 'LIKE', $request->search . '%')
+                ->orderBy('user.nama', 'asc')
+                ->paginate(10);
         } else {
-            $dataKaryawan = User::whereIn('jabatan', ['Konsultan', 'Project Manager', 'Staff', 'Support Manager'])
+            $dataKaryawan = User::join('role_has_user', 'user.id', '=', 'role_has_user.fk_user')
+                ->where('role_has_user.fk_role', 2)
                 ->orderBy('nama', 'asc')
-                ->paginate(20);
+                ->paginate(10);
         }
         return view('halaman_admin.karyawan.index', [
             'title' => $title,
@@ -40,13 +43,12 @@ class KaryawanController extends Controller
     {
         $request->validate([
             'email' => 'required|unique:user,email',
-            'signature' => 'required|unique:user,ttd'
         ], [
             'email.unique' => 'Email tidak boleh digunakan kedua kali!',
             'email.required' => 'Email tidak boleh kosong',
-            'signature.required' => 'Tanda Tangan tidak boleh kosong'
         ]);
 
+        // dd($request->all());
         $image_parts = explode(";base64,", $request->input('signature'));
         $image_type_aux = explode("image/", $image_parts[0]);
         $image_type = $image_type_aux[1];
@@ -55,7 +57,6 @@ class KaryawanController extends Controller
 
         $filePath = 'signatures/' . $filename;
         $fullFilePath = public_path($filePath);
-        // dd($filename);
         // Simpan file dalam direktori public
         file_put_contents($fullFilePath, $image_base64);
         $data = [
@@ -66,14 +67,23 @@ class KaryawanController extends Controller
             'no_rekening' => $request->no_rekening,
             'no_telp' => $request->no_telp,
             'bank' => $request->bank,
-            'ttd' => $filename
         ];
-        User::create($data);
+        if ($filename) {
+            $data['ttd'] = $filename;
+        } else {
+            $data['ttd'] = null;
+        }
+        $userNew = User::create($data);
+        Role_Has_User::create([
+            'fk_user' => $userNew->id,
+            'fk_role' => 2
+        ]);
         return redirect()->route('admin.karyawan')->with('success', 'Data Karyawan Berhasil Ditambahkan!');
     }
     public function edit_karyawan($id)
     {
-        $dataKaryawanEdit = DB::table('karyawan')->where('id', $id)->first();
+        $dataKaryawanEdit = User::find($id);
+        // dd($dataKaryawanEdit);
         $title = 'Edit Karyawan';
         return view('halaman_admin.karyawan.edit_karyawan', [
             'title' => $title,
@@ -82,18 +92,49 @@ class KaryawanController extends Controller
     }
     public function update_karyawan(Request $request, $id)
     {
-        DB::table('karyawan')->where('id', $id)->update([
+        // $image_parts = explode(";base64,", $request->input('signature'));
+        // $image_type_aux = explode("image/", $image_parts[0]);
+        // $image_type = $image_type_aux[1];
+        // $image_base64 = base64_decode($image_parts[1]);
+        // $filename = 'TTD_' . date('YmdHis') . '.' . $image_type;
+
+        // $filePath = 'signatures/' . $filename;
+        // $fullFilePath = public_path($filePath);
+        // // Simpan file dalam direktori public
+        // file_put_contents($fullFilePath, $image_base64);
+        $data = [
             'email' => $request->email,
             'nama' => $request->nama,
+            'password' => Hash::make($request->password),
             'jabatan' => $request->jabatan,
-            'password' => Hash::make($request->password)
-        ]);
-        return redirect()->route('admin.karyawan')->with('success', 'Data Karyawan Berhasil Diperbarui!');
+            'no_rekening' => $request->no_rekening,
+            'no_telp' => $request->no_telp,
+            'bank' => $request->bank,
+        ];
+        // if ($filename) {
+        //     $data['ttd'] = $filename;
+        // } else {
+        //     $data['ttd'] = null;
+        // }
+
+        $updatedKaryawan = User::find($id);
+        if ($updatedKaryawan) {
+            $updatedKaryawan->update($data);
+            return redirect()->route('admin.karyawan')->with('success', 'Data Karyawan Berhasil Diperbarui!');
+        } else {
+            return redirect()->route('admin.karyawan')->with('gagal', 'Data Karyawan Gagal Diperbarui!');
+        }
     }
     public function hapus_karyawan($id)
     {
-        $karyawan = Karyawan::findOrFail($id);
-        $karyawan->delete();
-        return redirect()->route('admin.karyawan')->with('success', 'Data Karyawan Berhasil Dihapus!');
+        $karyawan = User::findorFail($id);
+        if ($karyawan) {
+            $role_has_user = Role_Has_User::where('fk_user', $id);
+            $role_has_user->delete();
+            $karyawan->delete();
+            return redirect()->route('admin.karyawan')->with('success', 'Data Karyawan Berhasil Dihapus!');
+        } else {
+            return redirect()->route('admin.karyawan')->with('gagal', 'Data Karyawan Gagal Diperbarui!');
+        }
     }
 }
